I was trying to go for the effect seen in PrBoom+:;
TODO: When a player hits a wall, it moves the weapon to the center of the screen and disables bob, to appear as if the player is actually really jammed up against the wall (and not bobbing against it like he's trying to return some sort of interesting sexual pleasure to TEKWAL...lol). So, maybe this can be done.