Removed unwanted link to dm_state.h.
FIXME: This file results in two completely different behaviors between Visual Studio and MinGW. With MinGW/GCC, RGL_Wipe_Melt() works as intended, while under Visual Studio, the same function IS COMPLETELY BROKEN!!! All of the other WIPE effects work the same between versions, but for whatever reason we are encountering issues with screen wiping. I might know what the cause of this is, so I'm going to run some tests and see where it goes. For now, this needs to be marked.