Updated /eecrash/ from Eternity - renamed all to 'edgecrashreporter', fixed bugs, and expanded on the main crash window dialogues. Only works under Win32.
//TODO: Add the EDGEVERSTR to the crashreporter (use git_info for pre-release), automatically add the last
3 Lines from DEBUG.TXT into CRASHREPORT.txt.

Also fixed up the crash utility. For now, it _needs_ to be built seperately, and the resulting EXE added to the 3DGE
build directory. Eventually, this should be generated as a post-build thing, and the resultant EXE placed in EDGE's
build dir, not its own. For now, the application works great and provides even more details than ever, instead of
EDGE just crashing out without I_Erroring properly. Special thanks to Quasar for licensing it under the GPLv2.