Restore sound vars in CFG
This is only needed for Windows to properly set the Sound Mode to the correct values. Without these, sound is borked.

TODO: Make it so whatever sound mode EDGE/SDL decides to pick, it should immediately be written to the CFGDEF vars, just to be safe.

Also, setting the SDL_setenv to "directsound" re-enables 16-bit sound but we lose 32-bit sound, will leave it alone unless someone really wants it around.