Rollback Cinematic Code
I have renamed the new Cinematic code that utilizes SDL_Kitchensink/ffmpeg to -> i_ffmpeg, and restored the ROQ decoder (i_cinematic), and added i_ffmpeg to the source files. As of now, EDGE will only compile i_cinematic, unless the define USE_FFMPEG is defined - then it will enable the i_ffmpeg code. I did this primarily to make the Master builds stable again, and we are not 100% sure we are going to stick with Kitchensink/ffmpeg for video decoding.