SDL Video Fixes
Trying experimental stuff, which seems to give me good results (though needs wider testing):

First off, don't create renderer for video playback once already in game, which is what I_SetScreenSize was doing since it is only handled in I_StartupGraphics for video playback, and this was causing a *massive* framerate tank for Fullscreen.

Next, remove glClears from I_StartFrame(). In my tests, this fixes the small window on the lower left part of the screen when resolution or screensize changes take place. This was only added for the widescreen monitor modes, so for now there will be a white border in intermission screens. Though, this really minimizes the constant calling to clear the screen every single frame. This function can be reworked to be specific for the monitor changes without needing to impact every single frame. Needs further testing. Also removed clears and HUD_Reset in I_SetScreenSize.

Improved v_sync handling.