Build System: Collect and move libs to /deps
The following dependencies have been moved from root to deps: lzma, bzip2, and zlib.

Additionally, physfs has been added in to that subfolder as well. Changed a few lines in that makefile so it was decided to to have it added as submodule.

zlib was updated to 12.1.1 and MSVC users have this removed as a pre-compiled library (should always point to our own zlib).