ffmpeg: SDL_Kit Upgrade, Comment
Rewrote for newest SDL_Kit, and ffmpeg. Even though this "works", currently i_ffmpeg.cc does not actually "play" the movie, despite the engine recognizing the lump, opening the movie, printing info, and starting the stream (to test, compile with USE_FFMPEG and wrap that around any calls to "i_cinematic".
SDL2 clears the screen for one second (while the video should be playing in its entirety) before immediately skipping to the title screen.
This seems to be a problem with either the decoded video upload to OpenGL (SDL_GL_BindTexture), or in CIN_UpdateCinematic (via SDL_Renderer, or "my_rndrr", which is the only real "non-context" part of SDL_GL that manipulates our OGL subsystem). The ROQ player currently just writes/uploads the textures straight to OpenGL. Though, when this was initially created, it worked at one point. KIT_Player, SDL2, and the EDGE renderer all experienced crazy rewrites since then, to be fair.
My hunch currently rests with SDL_GL_BindTexture, or the SDL_Renderer code..