Bugfix for 'telept_fov' if damage > 0
This fixes a bug reported by Runsaber that locks the 'telept_fov' value (if "warp" is selected as the teleportation screen effect) if the player gets damaged at the same time as the effect takes place. I can probably do better in the future but this will work for now.