Feature: IWAD Selection
If EDGE does not find an IWAD, instead of erroring out, the user is able to specify both a directory and the IWAD itself. If the user does not specify one, then the engine will call I_Error.

This only needs to be done ONCE, since the variables are now correctly stored and configured in EDGE.INI. Which means after that selection is properly made, EDGE will continue to remember the IWAD locations. If you want to select a new dir and IWAD, the CVAR in the INI will have to be blown out manually, which will allow the user to save these settings.
Alternatively, this can also be set dynamically in an existing EDGE session via the Console. You can still continue to drag and drop EPKs, or PWADS.

This is the first stepping stone to a larger feature, which has been dormant for two decades: the Edge Game Module (.EGM) system that EDGE will eventually use.